+++ android/app/build.gradle
@@ -50,16 +50,11 @@ android {
     }
 
     defaultConfig {
-        // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
         applicationId "com.turnix.nylo"
-        // applicationId "com.turnix.nylo"
-        // applicationId "com.turnix.nylo.stg"
-        // You can update the following values to match your application needs.
-        // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
         minSdkVersion 21
         targetSdkVersion flutter.targetSdkVersion
         versionCode flutterVersionCode.toInteger()
-        versionName flutterVersionName
+        versionName "1.0.1"
     }
 
     signingConfigs {
