+++ lib/pages/home/tutor/functions/name_formatter.dart
@@ -1,0 +1,9 @@
+String formatName(String name) {
+  final String fullName = name;
+  final List<String> nameParts = fullName.split(' ');
+  final String firstName = nameParts[0];
+  final String format = firstName.substring(0, 1).toUpperCase() +
+      firstName.substring(1).toLowerCase();
+
+  return format;
+}
